feat(npc): support basic icache perf events

+++ npc/csrc/cpu/cpu_exec.c
@@ -18,7 +18,7 @@ extern char* perf_file;
 #ifdef CONFIG_PC_MAX_REPEAT
 static uint32_t pc_repeat_count = 0;
 static paddr_t  pc_old = 0;
-static uint32_t cycles = 0;
+static uint64_t cycles = 0;
 #endif
 
 void init_disasm();
@@ -227,7 +227,7 @@ void perf_event_display(){
     printf("perf_file: %s\n",perf_file);
     if (perf_file) fp = fopen(perf_file,"w");
     fprintf(fp,"******* Performance counter *******\n");
-    fprintf(fp,"Cycles: \t %.12d\n", cycles);
+    fprintf(fp,"Cycles: \t %ld\n", cycles);
     top_perf_event_display(fp);
     fprintf(fp,"******* End Performance counter *******\n");
     if (perf_file) fclose(fp);
